[puppet] Create Puppet 'profiles::bootstrap' class
The 'profiles::bootstrap' class serves as the "entrypoint" from which
*all* nodes (regardless of "type") invoke at the end of the Debian
preseed operation.

The purpose of the 'profiles::bootstrap' class is to perform the
minimum amount of operations required to prepare the node for a "full"
Puppet run on first boot.

The Debian preseed "chrooted" environment from which the
'profiles::bootstrap' class is invoked within is not a "true"
Debian GNU/Linux environment (e.g. 'systemd' init isn't running). A
result of this is that certain Puppet modules will not operate
correctly.

Unlike traditional "profiles" in the Puppet Roles & Profiles best-practice
pattern this "profile" serves as a "meta-profile" determining whether to
include the "seed" subclass for additional necessary configuration.

Initial class commit.