[puppet] Create Puppet 'profiles::bootstrap::seed' subclass
The 'profiles::bootstrap::seed' subclass performs the necessary
additional "seed" related configurations. It is intended to _only_ be
used by a singular "seed" node in the environment.

The purpose of this class is to create a static, single, YAML based fact
for reference by Hiera when performing its "Data-driven class
assignment". This extraneous, static fact assignment is only necessary
for the "seed" node to adhere to the singular "role" assignment when
Hiera operates as the External Node Classifier (ENC).

It should be noted that the `classes` encapsulated by the "seed" "role"
could have been listed in the "seed" node's `fqdn` lookup Hiera
data file. For consistency purposes this class permits the `classes` to
be encapsulated in the "seed" "role" Hiera data file:
`data/roles/seed.yaml`

By default the subclass's parent, 'profiles::bootstrap', only `include`s
the 'profiles::bootstrap::seed' when class parameter `seed` is
specified.