[hiera] Configure DHCP server for "raft" domain
The "raft" domain operates within a sizable, "Class B-style" IPv4 network
in which the "seed" node is responsible for managing by means of a DHCP
server (`isc-dhcp-server`). Future development will focus on isolating
further subnets to per-project/service (e.g. OpenStack).

Global DNS & NTP servers are leveraged at this early stage until
localised equivalents are implemented.

Tidbit: A group of penguins in the ocean are referred to as a "raft".