[puppet] Expose DNS "View(s)" capability
DNS server (BIND 9) "Views" permit the configuration of query
response behaviour depending on the query's source IP, destination IP,
type (e.g. AFXR), if it is recursive, and more.

The community module, 'theforeman-dns', supports the application of
"Views". This commit exposes the necessary interface for data-driven
creation of the 'dns::views' resource type.

Note: "Views" are *not* a standardised part of the DNS specification.
They are however widely implemented and should be actively leveraged
regardless of environment size.