[puppet] Create Puppet 'profiles::ssh' class
Operates as a minimal wrapper around the approved 'ghoneycutt-ssh'
component module.

By design the SSH profile class, 'profiles::ssh', configures the OpenSSH
server to accept public key authentication as the *only* mechanism for
authentication. This constraint requires the user to provide one or more
public key(s) and/or explicitly enable traditional password
authentication.

Explicitly disable the 'hiera_merge' 'ghoneycutt-ssh' component module
parameter that is intended to be set to `true` in future revisions. The
underlying `hiera_hash` function is deprecated and a Hiera v5
'lookup_options' approach is to be adopted instead.

Initial class commit.