[hiera] Reorder profiles::dns before profiles::dhcp
Puppet processes "Profiles" in the order in which they are declared in
their corresponding "Role" Hieradata file. ISC DHCP server's dependency
on the ISC BIND's generated RNDC key is satisfied by reordering the
application of the two "Profiles" as opposed to attempt to establish
dependency checking within either module.

While not as visible as explicitly defining the ordering requirements in
Puppet manifest DSL this approach does align with the Separation of
Concerns (SoC) principle.