[r10k] Remove r10k configuration file
The 'r10k.yaml' configuration file was incorrectly added as part of the
0.1.0 release (commit: 771ffbb0d5a377ce53f587cf1d80472f44ab361f). This
file was not referenced during `puppet agent` invocations and was
nonsensical within the context of the penguinspiral/puppet-control-repo
Git repository.

The _actual_ `r10k` global configuration file is stored in the
penguinspiral/puppet-r10k Git repository. The 'r10k.yaml' file is
referenced upon _each_ `puppet agent` invocation when the node is
operating in a "standalone" (i.e. locally running puppetserver)
deployment.