[puppet] Remove profiles::bootstrap::node subclass & meta-profile
With the removal of the 'profiles::bootstrap::seed' subclass (see:
9212e8f) the parent class, 'profiles::bootstrap', no longer needs to
determine whether to `include` it based on the 'seed' parameter.

The result of this eventual conditional removal means that the
'profiles::bootstrap' parent class operates as a redundant
abstraction/wrapper layer.
The ultimate intent here is to migrate 'profiles::bootstrap::node::{server,agent}'
sub-subclasses to subclasses directly under the 'profiles::bootstrap'
class. This can be represented as follows:

* Before: 'profiles::bootstrap::node::{server,agent}'

* After:  'profiles::bootstrap::{server,agent}'