[puppet] Remove profiles::bootstrap::seed subclass
By default the subclass's parent, 'profiles::bootstrap', only `include`s
the 'profiles::bootstrap::seed' when class parameter `seed` is
specified.

The purpose of this class was to create a static, single, YAML based fact
for reference by Hiera when performing its "Data-driven class
assignment".
Originally this extraneous, static fact assignment was only necessary
for the "seed" node to adhere to the singular "role" assignment when
Hiera operates as the External Node Classifier (ENC).

Extending this Hiera driven, statically derived ENC pattern to nodes of
*any* "type" (i.e. role) involves the removal of this special one-off
"role-to-class" implementation.
Instead the underlying implementation of creating a static "role"
identifier is to be handled by the 'ghoneycutt/facter' Puppet module
while the sourcing of the value of the targeted "role" is to be passed
as an environment variable: `FACTER_ROLE`

Adopting the Facter-readable environment variable provided a deployment
agnostic way that was consumable by both the "seed" node via the
customised Debian GNU/Linux installer and all other node "types" via
iPXE.
In both scenarios the kernel commandline contains the `FACTER_ROLE` key
with the corresponding targeted "role" value of which is sourced as a
runtime environment variable by the Debian GNU/Linux installer.