[rspec] Add custom 'role' RSpec fact
The custom 'role' "fact" is intended to be consumed via Facter in the
form of an environment variable. Commit 1ef030c outlines the rationale
for the 'role' fact and the manner in which it can be handled in various
deployment mechanisms.

For the purpose of RSpec tests this custom fact, 'role', has been set to
the value of 'seed' to simulate manifest generation on the initial
"seed" node in an environment.